factor out stuff from load_lora_adapter().
